Azure file upload works
